Add support for Prima.IgnoreEncodings resource.
Unix Prima spends a lot of time requesting information about various
fonts from the X server.  This leads to high startup times if the number
of fonts with similar names but different encodings is large, especially
for remote connections.  Meanwhile, the majority of these fonts is never
used.

To reduce the startup time, a new optional resource,
Prima.IgnoreEncodings, is introduced.

For example, on a system with approximately 7000 fonts, the startup time
of some application was 2.7 seconds.

By defining the following resource:

Prima.IgnoreEncodings:0;2;3;4;5;6;7;8;9;10;13;14;15;dectech;fontspecific;irv

the startup time of the same application was reduced to 1.15 seconds.

The same application, remote case:

No Prima.IgnoreEncodings: 171 seconds.
With Prima.IgnoreEncodings set as above: 64 seconds.


This solution is a hack.