layered widgets cannot have children
because of win32 passes mouse events by when alpha channel is not
set. One can try to insert f.ex. a button in a layered widget, but
it will be impossible to press it, because all GDI painting sets
alpha channel to 0 (even if there were non-0 values). This makes
support of layered children impractical. There's still difference
between x11 and win32 with mouse events passing by, and that's more
than enough.