extend and rename Drawable.fillWinding into .fillMode, with new constants
The reason is that the initial lofty promise to Prima to behave not only
identically on all platforms, but also in a sane way has met subtle problems
with filled shapes.  Namely, as filled shape do not occupy all area as defined
by vertices, one has to overlay it with a polygon to get a correct shape. This
is hard for cases with ROPs and fill patterns, and apparently sometimes is just
not worths it. So fillWinding is not fillMode( fm::Winding or fm::Alternate ),
plus a bitwise fm::Overlay that explicitly (and by default) tells to overlay
the polyline, but beware of side effects.