workaround Xorg/Xwayland garbled input bug with stippled output after repeated XSetRegion
I observed garbled input in a rather rare but 100% reproducible conditions when drawing
grids. The issue is rather complicated but it is truly a bug (will describe later)
that could be seen on my ubuntu 18. It's so complicated so I gave up to write a simple
perl script to show it (and even then a pure C version would be needed to submit as a bug
report, so I gave up). But the following needs to be true, apparently:

1) invalid region must contain 2 disjoint rectangles
2) paint must use XSetFillStyle(FillOpaqueStippled) drawing rectangle/bar over one of these rectangles
3) paint must then issue XSetRegion altering the region set in #1 in some way
4) paint must draw solid lines along the rectangles from #1

This gets the freshly drawn steps 3 and 4 blitted over with old pixels, leaving
only #2 parts visible on the screen. With #3 and #4 ignored, that drawing is
garbled horribly. I ran that with delays and could that rectangles between #1
and #2 were properly cleared, then #2 was drawn over, then #3 and #4, and then
suddenly the painted bits were completely reverted, except the bits from #2.

Looks very much a XWayland/wayland issue, but it's not there on my another ubuntu 18 laptop.