include hacked copy of libxrender.XRenderCompositeDoublePoly
In https://gitlab.freedesktop.org/xorg/lib/libxrender/-/issues/1 , Keith
Packard suggests not to use xrender tesselation as it is broken and coredumps.
I'll go with a crudely hacked copy of it though, until a more proper
tesselation code is found. So far it works on primitive cases and doesn't
coredump.