xrender and PictOpSrc messes up borders of a rotated image
When plotting rotated non-alpha images or rectangles, PictOpSrc results in a
strictly rectangular render, where the rotated shape appears correct but it
augmented with zero pixels. This is not happening with PictOpOver but that
means that in order to emulate SrcCopy with PictOpOver we need to explictly
address the source alpha.

Introduce another parameter to image caching where depending on the type of
image plotting, source alpha is either 0 (for standard X11 put_image) or 255
(for xrender-emulated SrcCopy)