Avoid excessive memory use caused by double rendering of arcs
Rewrite the path rendering so that polyline that is itself a result of
arc rendering will not be subject of eventual processing of line joins
that in case of lj::Round would cause lots of small arc to appear and
flood the final polygon with unneeded vertices. A simple ellipse with
diameter about 1000 pixels caused 200M+ memory use because of these
small arcs and the polygons they produced. Add hinting mechanisms to
.widen so that line joins could be overridden in certain parts of a
polygon .