refs #100: fixes for fvwm
as we're testing virtual size, not real size, the specificities
of WMs, namely how much they delay ConfigureNotify and how Prima
deals with it, is not really the point of this test.

Here, fvwm2 indeed delays these events and they come long after the last
restore() call, but it's not wrong in a sense that Prima cannot wait for
these forever; if a WM is slow, then it's slow.