unix/gtk.c: Make Prima's child process robust against signals
Signals sent to the application propagate to the whole process
group.  If a Prima application handles / ignores a signal, the
same signal can kill the child, which (correctly) discards any
signal handlers set by the application.

The patch puts the child in its own process group so that it
does no longer receive signals sent to the parent process.

This came up in a discussion on PDL: https://github.com/PDLPorters/pdl/pull/492